feat: Lexer has a readonly state
BREAKING CHANGE: the advance function of the lexer does not change the internal state of the object, but does instead return a new Lexer object. The getters `token()`, `peek()` and `previous()` have been replaced with readonly properties `current`, `next`, `previous`.

BREAKING CHANGE: the API of the Parser class has changed. It can't parse multiple expressions anymore, but needs to be instantiated for each expression. The expression is passed to the constructor. The current lexer can be accessed via the property `lexer`.